+++ Sources/KeyboardKit/Callouts/EnglishSecondaryCalloutActionProvider.swift
@@ -31,6 +31,7 @@ open class EnglishSecondaryCalloutActionProvider: BaseSecondaryCalloutActionProv
             
         case "-": return "-–—•"
         case "/": return "/\\"
+        case "$": return "$€£¥₩₽¢"
         case "&": return "&§"
         case "”", "“": return "\"”“„»«"
         case ".": return ".…"